pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-scraper"
version = "0.1.0"
description = "Pokemon Showdown Replay Scraper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.scripts]
pokemon-scraper = "showdown_scraper.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["showdown_scraper*"]
